requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
stripe==7.6.0
python-dotenv==1.0.0
Werkzeug==3.0.1
WTForms==3.1.1
requests
flask
flask-login
flask-sqlalchemy
flask-wtf
psycopg2-binary
python-dotenv
stripe
werkzeug
wtforms
